Move PollCreatePayload to DSharpPlus.Net.Abstractions
