+++ DSharpPlus/Entities/BaseDiscordMessageBuilder.cs
@@ -3,8 +3,8 @@ using System.Collections.Generic;
 using System.IO;
 using System.Linq;
 using System.Threading.Tasks;
-using DSharpPlus.Entities.Channel;
 using DSharpPlus.Net;
+using DSharpPlus.Net.Abstractions;
 
 namespace DSharpPlus.Entities;
