+++ DSharpPlus/Net/Abstractions/Rest/RestChannelPayloads.cs
@@ -1,7 +1,6 @@
 using System;
 using System.Collections.Generic;
 using DSharpPlus.Entities;
-using DSharpPlus.Entities.Channel;
 using Newtonsoft.Json;
 
 namespace DSharpPlus.Net.Abstractions;
